requirements.txt
hypercorn==0.14.4
Flask==2.2.5
requests==2.31.0
replicate==0.25.0
sqlalchemy==2.0.28
openai==1.25.0
boto3==1.34.101
gunicorn==20.0.4
itsdangerous
MarkupSafe
Werkzeug
starlette
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
flask_restplus
pymysql==1.1.1